+++ CHANGES.md
@@ -8,6 +8,10 @@
 
 * Added `URLPath` option for logger
 
+* Added `ServiceRequest::app_data()`, returns `Data<T>`
+
+* Added `ServiceFromRequest::app_data()`, returns `Data<T>`
+
 ### Changed
 
 * Move multipart support to actix-multipart crate